feat: learn how to compile TS to JS (#2)
* feat: output files to dist directory

Output files are now placed in a "dist" directory instead of the root directory. This change allows for a cleaner separation of source files and output files.

* fix: add dist to gitignore

This commit fixes an issue where the dist directory was not being ignored by git.

BREAKING CHANGE: The dist directory is now ignored by git. This may cause issues if you are relying on the dist directory to be included in your git commits.

* feat: improve tsconfig.json

Add "include" and "exclude" options to tsconfig.json to improve code organization.