feat: learn interface implementation in TS (#12)
* feat: implement data type interface

Adds a new data type interface for sellers and updates tests to use the interface.

* feat: update data type interface to readonly id

This commit updates the data type interface to make the id property readonly. This prevents the id from being accidentally changed, which could lead to unintended consequences.

BREAKING CHANGE: The id property of the data type interface is now readonly. This means that it cannot be changed after the object is created.