feat: learn indexable interface (#14)
* feat: introduce indexable interface

Introduces an indexable interface for arrays of strings.

This allows for easier access to elements in an array by their index.

* feat(indexable-interface): add StringMap indexable interface

This commit introduces a new `StringMap` indexable interface to the `indexable-interface` package.

The `StringMap` interface provides a way to index strings by strings, allowing for more flexible and expressive data structures.

This interface can be used to represent a variety of data structures, such as:

- Key-value pairs
- Maps
- Dictionaries
- Objects

The `StringMap` interface is a valuable addition to the `indexable-interface` package, and it provides a number of benefits for developers who are working with strings in JavaScript.