feat(extending-interface): add support for extending interface
Adds support for extending interfaces. This allows us to create new interfaces that inherit properties and methods from existing interfaces.

This feature is useful for creating hierarchies of objects that share common properties and methods, but also have unique properties and methods. For example, we could create an Employee interface with properties like name and id, and then create a Manager interface that extends the Employee interface and adds properties like numOfEmployees.