feat: add function in interface
Adds a new feature that allows functions to be defined in interfaces.

This allows for more flexible and extensible code, as functions can now be passed around as objects and used in a variety of ways.

BREAKING CHANGE: The syntax for defining functions in interfaces has changed. Functions must now be defined using the `function` keyword, followed by the function name and parameters.