feat: support intersection type
Adds support for intersection types.

Intersection types allow us to create new types from existing types.
This can be useful for creating types that represent objects that have multiple different properties.
For example, we can create a type that represents an object that has both an ID and a name:

```typescript
type Domain = HasID & HasName;
```

This type would require its instances to have both an id and a name.