feat: add support for rest parameter
This commit adds support for the ES6 rest parameter syntax in the `sum` function.

This allows the function to accept an arbitrary number of arguments, which are then collected into an array and summed.

This is useful for cases where you need to sum a variable number of values, such as when you are calculating the total value of a shopping cart or the average of a set of numbers.