feat: learn function in TS (#19)
* feat: implement function support

Add support for function types, including arrow functions, named functions, and void functions.

BREAKING CHANGE: Functions now return values instead of printing to the console.

* feat: make function parameters optional

This commit makes function parameters optional by default.

This allows users to call functions without providing arguments, which can be useful in situations where the default values are appropriate.

Additionally, this change simplifies the API of the function, making it easier to use.

* feat: add support for rest parameter

This commit adds support for the ES6 rest parameter syntax in the `sum` function.

This allows the function to accept an arbitrary number of arguments, which are then collected into an array and summed.

This is useful for cases where you need to sum a variable number of values, such as when you are calculating the total value of a shopping cart or the average of a set of numbers.

* feat: support optional parameter in sayHello

Adds an optional lastName parameter to the sayHello function.

This allows the function to be more flexible, as it can now be used to greet people with or without a last name.