feat: learn loop statement in TS (#25)
* feat(test): add loop test

Adds a test suite for the loop function, covering various loop syntaxes including for with manual index, for with element, for with index, and forEach.

* test: add while loop test

Adds a test for the while loop.

* feat: add do while loop

Add a function to test the do while loop.

* feat(loop): implement break and continue

Adds the ability to use break and continue statements within loops.